% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{barcodeLabel}
\alias{barcodeLabel}
\title{barcodeLabel}
\description{
A simple R package for fast QR code generation and barcode label sheet design (such as Avery label sheets).
}
\author{
Junli Zhang <zhjl86@gmail.com>
}
