% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLabel.R
\name{code_128_make}
\alias{code_128_make}
\title{code_128_make: create a grid Grob object for drawing linear code128}
\usage{
code_128_make(Labels)
}
\description{
code_128_make: create a grid Grob object for drawing linear code128
}
