% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrcode.R
\name{dmcode}
\alias{dmcode}
\title{Make data matrix code}
\usage{
dmcode(text, square = 1)
}
\arguments{
\item{text}{input text for datamatrix code generation}

\item{square}{whether to use squared shape (1 is Yes; 0 is NO). If not, the function will automatically decide the shape.}
}
\value{
a number matrix representing the datamatrix code
}
\description{
Make data matrix code
}
\examples{
dmout <- dmcode("Hello, world!", 1)
plotqr(dmout)
}
