% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLabel.R
\name{dmcode_make}
\alias{dmcode_make}
\title{dmcode_make: create a grid Grob object for drawing data matrix code generated by function "dmcode"}
\usage{
dmcode_make(Labels, square = 1, scale = 1)
}
\arguments{
\item{scale}{0-1 barcode scaling inside the drawing area}
}
\description{
dmcode_make: create a grid Grob object for drawing data matrix code generated by function "dmcode"
}
