% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrcode.R
\name{plotqr}
\alias{plotqr}
\title{Plot the QR code}
\usage{
plotqr(x)
}
\arguments{
\item{x}{output of qrcode function}
}
\description{
Plot the QR code
}
\examples{
qrout <- qrcode("Hello, world!", 1)
plotqr(qrout)
}
