% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrcode.R
\name{qrcode}
\alias{qrcode}
\title{A function to make QR code raw data for plotting. This function is very fast because it uses C code in the hood.}
\usage{
qrcode(text, ecl = 1)
}
\arguments{
\item{text}{input text for QR code generation}

\item{ecl}{Error correction level (1 = Low (tolerate about  7% erroneous codewords); 2 = Medium (15%); 3 = QUARTILE (25%); 4 = High (30%)). 
The correction level of the result may be higher than the ecl argument if it can be done without increasing the QR code version (dimension).}
}
\value{
a number matrix representing the QR code
}
\description{
A function to make QR code raw data for plotting. This function is very fast because it uses C code in the hood.
}
\examples{
qrout <- qrcode("Hello, world!", 1)
plot(qrout)
}
