% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLabel.R
\name{qrcode_make}
\alias{qrcode_make}
\title{qrcode_make: create a grid Grob object for drawing QR code generated by function "qr"}
\usage{
qrcode_make(Labels, ErrCorr = 2, scale = 1)
}
\description{
qrcode_make: create a grid Grob object for drawing QR code generated by function "qr"
}
