% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richtext.R
\name{richtext}
\alias{richtext}
\title{Draw formatted text labels
richtext() can draw text labels with mixed fontface from markdown syntax (only support * quotes)" \*italic\*, or \*\*bold\*\*, or \*\*\*bold italic\*\*\*}
\usage{
richtext(
  txt,
  x = 0.5,
  y = 0.5,
  gp = grid::gpar(),
  unit = "npc",
  hjust = 0.5,
  vjust = 0.5,
  draw = TRUE,
  useMarkdown = TRUE
)
}
\arguments{
\item{txt}{a string}

\item{x}{coordinate on the x axis}

\item{y}{coordinate on the y axis}

\item{gp}{An object of class "gpar", typically the output from a call to the function gpar.}

\item{unit}{A string indicating the default units to use if x or y are only given as numeric values.}

\item{hjust}{A number specifying horizontal justification.}

\item{vjust}{A number specifying vertical justification.}

\item{draw}{A logical value indicating whether graphics output should be produced.}

\item{useMarkdown}{whether treat ** quotes as markdown (only support fontfaces)}
}
\value{
richtext() returns a gList that can be used in grid::grid.draw
}
\description{
Draw formatted text labels
richtext() can draw text labels with mixed fontface from markdown syntax (only support * quotes)" \*italic\*, or \*\*bold\*\*, or \*\*\*bold italic\*\*\*
}
\examples{
library(grid)
grid.newpage()
vp = viewport(x=0.5, y=0.5, width=unit(3.5, "inches"), height=unit(1, "inches"), gp=gpar(fontsize=12, lineheight=0.8))
pushViewport(vp)
grid.rect()
# can use "\n" for a new line
txt = "plain text and\n *italic text* and **bold text** and\n***bold italic text***"
gg = richtext(txt, draw = TRUE)
# grid.draw(gg)
popViewport(1)

}
