% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_wrap.R
\name{text_box_wrap}
\alias{text_box_wrap}
\title{Wrap text within a text box}
\usage{
text_box_wrap(text, font_size, box_width, box_height, fontfamily = "sans")
}
\arguments{
\item{text}{a single string}

\item{font_size}{font size}

\item{box_width}{text box width in inch}

\item{box_height}{text box height in inch}

\item{fontfamily}{font family ("mono", "sans", "serif") or specific font based on your operation system.}
}
\value{
a vector with 2 elements: new string with "\n" added and the final adjusted font size
}
\description{
Wrap text within a text box
}
\examples{
box_width = 1
box_height = 0.5
text = "line1\na-long-text-line_long_long_long"
font_size = 10
text_box_wrap(text, font_size, box_width, box_height, "mono")
}
