% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_wrap.R
\name{word_spit}
\alias{word_spit}
\title{Split string to word with non-alpha-numeric characters}
\usage{
word_spit(str)
}
\arguments{
\item{str}{a string}
}
\value{
a vector of split 'words'
}
\description{
Split string to word with non-alpha-numeric characters
}
\examples{
word_spit("abc\n123 abc-ef-ghijk*ll")
}
