% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_wrap.R
\name{str_wrap_inch}
\alias{str_wrap_inch}
\title{A function to wrap string for width (inch) limit}
\usage{
str_wrap_inch(str_vec, width)
}
\arguments{
\item{str_vec}{a vector of split words vector from word_split}

\item{width}{width limit in inch}
}
\value{
a string with "\n" added between words. You can use `cat` to see the word wrap
}
\description{
A function to wrap string for width (inch) limit
}
\examples{
text = "line1\na-long-text-line_long_long_long"
str_vec = word_spit(text)
new_text = str_wrap_inch(str_vec, width = 1.2)
cat(new_text)
}
