% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_wrap.R
\name{text_array_wrap}
\alias{text_array_wrap}
\title{Wrap a vector of strings to fit a text box.
Just a vector function of "text_box_wrap"}
\usage{
text_array_wrap(
  text_array,
  font_size = 12,
  box_width,
  box_height,
  fontfamily = "sans"
)
}
\arguments{
\item{text_array}{a string vector}

\item{font_size}{inital font size}

\item{box_width}{text box width in inch}

\item{box_height}{text box height in inch}

\item{fontfamily}{font family ("mono", "sans", "serif") or specific font based on your operation system.}
}
\value{
a list of 2 elements: a vector of wrapped string and the final font size.
}
\description{
Wrap a vector of strings to fit a text box.
Just a vector function of "text_box_wrap"
}
\examples{
ss = c("a very long long string", "another-long-long-string")
text_array_wrap(ss, 12, 1, 0.5)
}
